First sweep of cleaning/clarifying the code
