+++ GSA_Adapter/IAdapter/Execute.cs
@@ -18,14 +18,5 @@ namespace BH.Adapter.GSA
             return ComCall(command);
         }
 
-
-        /***************************************************/
-        /**** Public  Methods                           ****/
-        /***************************************************/
-
-
-        /***************************************************/
-        /**** Private  Methods                          ****/
-        /***************************************************/
     }
 }
