+++ GSA_Adapter/IndexAdapter/Comparer.cs
@@ -1,0 +1,46 @@
+﻿using BH.oM.Materials;
+using BH.oM.Structural.Elements;
+using BH.oM.Structural.Properties;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace BH.Adapter.GSA
+{
+    public partial class GSAAdapter
+    {
+        /***************************************************/
+        /**** BHoM Adapter Interface                    ****/
+        /***************************************************/
+
+        protected override IEqualityComparer<T> GetComparer<T>()
+        {
+            Type type = typeof(T);
+
+            if (m_Comparers.ContainsKey(type))
+            {
+                return m_Comparers[type] as IEqualityComparer<T>;
+            }
+            else
+            {
+                return EqualityComparer<T>.Default;
+            }
+            
+        }
+
+
+        /***************************************************/
+        /**** Private Fields                            ****/
+        /***************************************************/
+
+        private static Dictionary<Type, object> m_Comparers = new Dictionary<Type, object>
+        {
+            {typeof(Node), new BH.Engine.Structure.NodeDistanceComparer(3) },
+            {typeof(SectionProperty), new BH.Engine.Base.BHoMObjectNameOrToStringComparer() },
+            {typeof(Material), new BH.Engine.Base.BHoMObjectNameComparer() },
+        };
+
+    }
+}
