+++ GSA_Adapter/IndexAdapter/UpdateTags.cs
@@ -17,14 +17,5 @@ namespace BH.Adapter.GSA
             return Create(objects);
         }
 
-
-        /***************************************************/
-        /**** Public  Methods                           ****/
-        /***************************************************/
-
-
-        /***************************************************/
-        /**** Private  Methods                          ****/
-        /***************************************************/
     }
 }
